+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### v1.3.4 - 2024 Apr.10
+- Auto Preview **Resolution**
+
+### v1.3.3 - 2024 Apr.10
+- Add Row **Colors**
+
+### v1.3.2 - 2024 Apr.08
+- Implement **Click & Drag** for Advanced Mapping
+- Auto **Preview**
+- Improved **Error Log**
+
+### v1.3.1 - 2024 Apr.08
+- Add **Row-Control Buttons** for Advanced Mapping
+
 ### v1.3.0 - 2024 Mar.30
 - Add **Advanced Mapping** *(manual regions)*